Add the infrastructure needed to generate a crude "OS" ISO.
Generate an ISO, cosmo.iso, containing the bootloader (GRUB) and a primitive
"OS" which calls a main() C function and then proceeds to loop forever. Using
the included Bochs configuration file, the ISO can be loaded and debugged
in Bochs. You should expect to see the value 0xDEADBEEF in register eax
(this is the output of running main()) indicating the "OS" ran successfully.