CMake-ify the project.
This commit has the cosmo project use CMake for build generation rather
than directly maintaining a Makefile. Part of this migration to CMake had
us restructure the project to include the following directories:

  * kernel  -> Contains kernel sources.
  * src     -> Contains all *.cc implementation files.
  * include -> Contains all *.h header files.
  * scripts -> Contains project scripts.
  * docs    -> Contains a CMakeLists.txt describing how to build project docs
               using Doxygen.

This commit leaves the project in a state where the kernel ELF and project docs
can be built/installed. A future commit will add a series of scripts to support
the build process as well as running the OS through Bochs.