Create a proper grub.cfg for the OS.
Previously, we were using a legacy version of the GRUB stage 2 bootloader
(stage2_eltorito). This commit reworks our iso/ directory such that
the stage 2 bootloader executable goes away and a new grub.cfg configuration
file replaces the menu.lst. The end result is that when we build and install
cosmo, we can genrate the ISO using grub-mkrescue. When booting the ISO
we now get a nice menu to select cosmo and execution proceeds as before. Note,
this change adds the following project dependencies:
    * GRUB (grub-mkrescue)
    * xorriso
    * GNU mtools