Disable warning about unused values when logging a fmt string with no args.
When using a Logger object to log a string with no fmt args (e.g.,
logger.LogInfo(writer, "Hello, world.")), the build fails because of an
unused value warning. The compiler correctly identifes that the
arg_array in the Log() method goes unused. It is safe to disable this
warning. To keep dependent projects using -Werror from breaking, the
-Wno-unused-variable compiler option is exported with INTERFACE in
src/Logger/CMakeLists.txt.