Update the Logger class to expose LOG_* convenience macros.
As we move forward with writing more complicated code that requires
debug through use of log messages, the Logger interface has become
particularly cumbersome. This revision, introduces LOG_* macros similar
to LOG4CXX macros that allow us to log messages without having to
specify the writer object everytime. This is especially useful when
logging to the framebuffer because the LOG_* macros take care of
fetching the singleton FrameBuffer object for us.