Implement a physical page frame allocator.
This revision introduces the PhysicalFrameAllocator class which is a
singleton class used to alloc/free 4KB page frames. Internally, used/unused
frames are tracked using a bitmap. The allocator places its bookkeeping
information at the end of the kernel space in memory. The majority of
this implementation was lifted from:
https://medium.com/lucians-blog/writing-a-physical-memory-manager-866fc1457d6