Make the artefacts generated by the container be owned by the host's user.
When using the development container to build cosmo, the binaries, logs,
etc. generated by the build process would be created under the dev
container's root user. On container exit, you would need sudo/root to clear
the files on the host. To remedy this, the relevant scripts and the dev
container has been updated so that the host system's user is automatically
the owner of all build artefacts.